Move draw crop code to new file
