+++ rtgui/previewwindow.cc
@@ -19,6 +19,7 @@
 #include "previewwindow.h"
 
 #include "cursormanager.h"
+#include "drawcropguide.h"
 #include "guiutils.h"
 #include "hidpi.h"
 #include "imagearea.h"
